Do not use 'class'
Older browsers don't support class, and for that reason, React components cannot be used - only React stateless elements.

Because we publish this app automatically to GitHub Pages, as soon as a commit is merged into the master branch, transpiling is not an easy option.

For these reasons, the ChipCollection is now stateless and is expanded or collapsed with jQuery.